0.9.64 - Метки времени на таймлайне (в снятом мульте); В метках времени убраны нули старшей размерности, мне они показались лишними. А номер кадра в пределах секунды заключён в квадратные скобки; Добавлено обновление строки состояния при включении двойной частоты кадров; Перетаскивание всех выделенных кадров в режиме управления кадрами, вместо перетаскивания одного того, что в фокусе; Строчка о  назначении программы в "О программе"; Рекомендация кликнуть на экран на заставке; Более адеватная реакция на удаление файла grab в работе окна захвата картинок с камеры;

+++ Sources/CameraFormUnit.pas
@@ -411,7 +411,9 @@ begin
   if Assigned(FFileCommanderDirMonitor) then
     for i := 0 to FFileCommanderDirMonitor.Notifications.Count - 1 do
       if (LowerCase(FFileCommanderDirMonitor.Notifications[i]) = 'grab') then
-        if ([dmaAdded, dmaNewName] * FFileCommanderDirMonitor.Notifications.Actions[i] <> []) then
+        if ([dmaAdded, dmaNewName] * FFileCommanderDirMonitor.Notifications.Actions[i] <> []) and
+           ([dmaRemoved] * FFileCommanderDirMonitor.Notifications.Actions[i] = [])
+        then
           try
             SavedFileName := GetFileContent(PhotoFolder + 'grab');
             DeleteFile(PhotoFolder + 'grab'); // забрали исполнять
